Create 2021-9-10[PullAndBranch].html
	Branches, in GitHub, is a way for people to keep current, finalized code, separate from developmental code that isn’t finished yet. With each repo in GitHub there is a main branch: the default branch. The default branch is where the finalized code is pushed to and for all people who have access to the repo, to see. The other branches you create are areas of the repo that you can test code without fear of it interfering with the main code. 

1)	How to create Branches

	First, to create a branch, you need write access to the repo. To create a new branch on GitHub desktop, you should see in your action bar at the top of the application, a dropdown that says “current branch: main.” With this dropdown you should see a button that says “New branch.” This button allows you to create a new branch to work under. With this dropdown menu you can also swap in between branches you’ve either created or the main branch.	

A Pull Request is a way for contributors of an open-source project to be looked at by someone who has, essentially, administrative access (also known as “maintainers”) to the repo. This allows maintainers to overlook proposed work without having it mess up existing code. If maintainers see the requested code and think it should be added to the existing, finalized code, they can merge the pull requested code. If they think the code shouldn’t be added they can decline the pull request and even leave comments over the proposed code.

2)	How to create Pull Requests

In the GitHub desktop application at the top of the window, you will see a dropdown menu selection called “Repository” through this dropdown you should see the option “Pull.” This option is the option that allows you to create a Pull Request for the maintainers to see.